rostful/requirements.txt
--index-url https://pypi.python.org/simple/

Babel==1.3
Flask==0.10.1
Flask-Celery-Helper==1.1.0
Flask-Cors==2.0.1
-e git://github.com/asmodehn/flask-restful.git#egg=Flask_RESTful-asmodehn
Flask-Redis-Helper==1.0.0
Flask-Script==2.0.5
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.10.4
amqp==1.4.6
aniso8601==1.0.0
anyjson==0.3.3
backports.ssl-match-hostname==3.4.0.2
billiard==3.3.0.20
celery==3.1.18
certifi==2015.04.28
flower==0.8.2
futures==3.0.2
gunicorn==19.3.0
itsdangerous==0.24
kombu==3.0.26
pytz==2015.4
redis==2.10.3
six==1.9.0
tornado==4.2

-e .